+++ serialization/src/main/java/org/finos/rune/mapper/RuneJacksonJsonObjectMapper.java
@@ -36,7 +36,7 @@ import com.fasterxml.jackson.module.paramnames.ParameterNamesModule;
 import com.rosetta.model.lib.RosettaModelObject;
 import org.finos.rune.mapper.date.RuneDateModule;
 import org.finos.rune.mapper.filters.SubTypeFilter;
-import org.finos.rune.mapper.json.RuneJSONModule;
+import org.finos.rune.mapper.introspector.RuneJSONModule;
 import org.finos.rune.mapper.mixins.RosettaModelObjectMixin;
 
 /**
