+++ serialization/src/main/java/org/finos/rune/mapper/introspector/RuneJSONModule.java
@@ -1,4 +1,4 @@
-package org.finos.rune.mapper.json;
+package org.finos.rune.mapper.introspector;
 
 /*-
  * ==============
